refactor(docs): trim reference validation scaffolding
Keep CLI reference parity in one package-local integration test with its
parser inlined, and remove redundant assertions and duplicate README
option coverage.

Drop the extra search/LLM HTTP smoke and its command and CI wiring. Keep
the existing docs build, source-derived tables, typechecked examples,
and package-triggered validation.

Remove the historical audit from the working tree and shorten checking
instructions and process commentary in the docs.
