+++ apps/docs/content/docs/guide/build.mdx
@@ -40,7 +40,7 @@ crust build --env-file .env.production
 
 ## Flags
 
-The mechanical columns below are parity-checked against the command definitions. “Declared default” is only a definition default; `—` means none is declared. Effective defaults are resolved later: runtime uses the flag, then `package.json`, then Bun; name is [auto-detected](#binary-naming); omitted targets select all platforms for Bun/Deno; minification defaults to enabled for Bun/Node and is unavailable for Deno.
+“Declared default” is only a definition default; `—` means none is declared. Effective defaults are resolved later: runtime uses the flag, then `package.json`, then Bun; name is [auto-detected](#binary-naming); omitted targets select all platforms for Bun/Deno; minification defaults to enabled for Bun/Node and is unavailable for Deno.
 
 | Flag spellings                | Type                  | Choices               | Declared default | Description                                                                         |
 | ----------------------------- | --------------------- | --------------------- | ---------------- | ----------------------------------------------------------------------------------- |
