+++ apps/docs/content/docs/modules/create-crust.mdx
@@ -36,7 +36,7 @@ create-crust [directory] [--distribution binary|runtime] [--install|--no-install
 | `--git`, `--no-git`             | `boolean` | —                   | —                |
 | `--overwrite`, `--no-overwrite` | `boolean` | —                   | —                |
 
-These option columns are parity-checked against the built CLI snapshot. No flag default is declared; the following are **prompt defaults and behavior**, not definition defaults:
+No flag default is declared; the following are **prompt defaults and behavior**, not definition defaults:
 
 - `--distribution`: Preselect the distribution prompt. The default is `binary` (standalone executables); `runtime` produces a package that runs with Bun.
 - `--install` / `--no-install`: Set the initial answer to the dependency-installation prompt. The default is to install dependencies.
