+++ apps/docs/package.json
@@ -7,8 +7,7 @@
     "dev": "vite dev",
     "build": "vite build",
     "preview": "vite preview",
-    "check:references": "cd ../.. && bunx turbo run build --filter=create-crust && bun test apps/docs/scripts/cli-reference.test.ts",
-    "check:output": "bun test tests/compiled-docs.test.ts",
+    "check:references": "cd ../.. && bunx turbo run build --filter=create-crust && bun test apps/docs/tests/cli-reference.test.ts",
     "check:types": "fumadocs-mdx && tsc --noEmit",
     "postinstall": "fumadocs-mdx",
     "deploy": "bun run build && wrangler deploy",
